// src/config.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Usernames = Username[];

export interface Config {
  usernames?: Usernames;
  [k: string]: unknown;
}
export interface Username {
  id: string;
  username: string;
  [k: string]: unknown;
}